KupData: cleaned helper.
